issues-1268: BaseArtifactController improvements
