+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/BaseArtifactController.java
@@ -11,66 +11,22 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.InputStream;
 
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-@RequestMapping(path = "/storages")
 public abstract class BaseArtifactController
         extends BaseController
 {
     @Inject
     protected ArtifactManagementService artifactManagementService;
 
-
-    @ApiOperation(value = "Used to deploy an artifact")
-    @ApiResponses(value = { @ApiResponse(code = 200, message = "The artifact was deployed successfully."),
-                            @ApiResponse(code = 400, message = "An error occurred.") })
-    @PreAuthorize("hasAuthority('ARTIFACTS_DEPLOY')")
-    @PutMapping(value = "{storageId}/{repositoryId}/{path:.+}")
-    public ResponseEntity upload(@ApiParam(value = "The storageId", required = true)
-                                 @PathVariable(name = "storageId") String storageId,
-                                 @ApiParam(value = "The repositoryId", required = true)
-                                 @PathVariable(name = "repositoryId") String repositoryId,
-                                 @PathVariable String path,
-                                 HttpServletRequest request)
-    {
-        try
-        {
-            RepositoryPath repositoryPath = repositoryPathResolver.resolve(storageId, repositoryId, path);
-
-            if (!repositoryPath.getRepository().isInService())
-            {
-                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Repository is not in service.");
-            }
-
-            artifactManagementService.validateAndStore(repositoryPath, request.getInputStream());
-
-            return ResponseEntity.ok("The artifact was deployed successfully.");
-        }
-        catch (Exception e)
-        {
-            logger.error(e.getMessage(), e);
-
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-        }
-    }
-
-    public Storage getStorage(String storageId)
+    protected Storage getStorage(String storageId)
     {
         return configurationManager.getConfiguration().getStorage(storageId);
     }
 
-    public Repository getRepository(String storageId,
+    protected Repository getRepository(String storageId,
                                     String repositoryId)
     {
         return getStorage(storageId).getRepository(repositoryId);
