+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/layout/maven/MavenArtifactController.java
@@ -116,6 +116,39 @@ public class MavenArtifactController
         provideArtifactDownloadResponse(request, response, httpHeaders, repositoryPath);
     }
 
+    @ApiOperation(value = "Used to deploy an artifact")
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "The artifact was deployed successfully."),
+                            @ApiResponse(code = 400, message = "An error occurred.") })
+    @PreAuthorize("hasAuthority('ARTIFACTS_DEPLOY')")
+    @PutMapping(value = "{storageId}/{repositoryId}/{path:.+}")
+    public ResponseEntity upload(@ApiParam(value = "The storageId", required = true)
+                                 @PathVariable(name = "storageId") String storageId,
+                                 @ApiParam(value = "The repositoryId", required = true)
+                                 @PathVariable(name = "repositoryId") String repositoryId,
+                                 @PathVariable String path,
+                                 HttpServletRequest request)
+    {
+        try
+        {
+            RepositoryPath repositoryPath = repositoryPathResolver.resolve(storageId, repositoryId, path);
+
+            if (!repositoryPath.getRepository().isInService())
+            {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Repository is not in service.");
+            }
+
+            artifactManagementService.validateAndStore(repositoryPath, request.getInputStream());
+
+            return ResponseEntity.ok("The artifact was deployed successfully.");
+        }
+        catch (Exception e)
+        {
+            logger.error(e.getMessage(), e);
+
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+        }
+    }
+    
     @ApiOperation(value = "Copies a path from one repository to another.")
     @ApiResponses(value = { @ApiResponse(code = 200, message = "The path was copied successfully."),
                             @ApiResponse(code = 400, message = "Bad request."),
